[WeatherApp] LocationService added and will fetch the current location weather info

+++ app/src/main/java/com/example/weatherapp/model/City.kt
@@ -1,3 +1,3 @@
 package com.example.weatherapp.model
 
-data class City(val name: String, val lat: Float, val lon: Float, val state: String, val country: String)
+data class City(val name: String, val lat: Double, val lon: Double, val state: String, val country: String)
